requirements.txt
dj-database-url==0.4.0
Django==1.9.2
django-cors-headers==1.2.2
djangorestframework==3.4.7
gunicorn==19.4.5
psycopg2==2.6.1
wheel==0.24.0
whitenoise==2.0.6
